termion: Always use Reset to clear formatting
Previously, we used the NoBold, NoItalic and NoUnderline commands to
clear effects, and we did not clear the colors.  With this patch, we
always use Reset to clear the formatting.  This fixes two issues:
- Not all ANSI terminals support the No Bold style.
- If we don’t reset the color and the next string does not have a color
  set, it will be printed in the wrong color.
Also, this is consistent with ansi_term’s and crossterm’s implementation
and therefore easier to test.