Refactor syntect example
This patch refactors the syntect example.  The main change is that we
now explicitly require that all features are activated and that we drop
all features checks (that didn’t work as intended anyway).  We also add
an introductory comment that explains what the example does and how to
run it.