Disable default features for syntect
We don’t have to pull in any syntect features for compiling the library.
Even for the syntect example, we don’t need all default features.  With
this patch, we disable the default features for the syntect dependency
and add a dev dependency with the features required for the example.