Disable default features for cursive
Previously, we pulled in the ncurses-backend for cursive for the syntect
example although we were using the crossterm backend.  With this patch,
we disbale the default features for the cursive dev dependency.  (It was
already disabled for the library dependency.)

Also, we change the dev dependency version to a wildcard because we
always want to match the version of the library dependency.